+++ src/HeadlessTree/VirtualizedTree.tsx
@@ -8,7 +8,7 @@ import {
   type ReactElement,
   type Ref,
 } from 'react';
-import { flattenTree } from './flattenTree';
+import { flattenTree } from './utils/flattenTree';
 import type { BasicTreeItem, TreeProps } from './types';
 import { useTreeState } from './useTreeState';
