+++ src/HeadlessTree/index.ts
@@ -1,8 +1,9 @@
-export { flattenTree } from './flattenTree';
 export { Tree } from './Tree';
 export type { TreeRef } from './Tree';
 export type * from './types';
 export { useTreeState } from './useTreeState';
-export { getAllDescendantIds, getPath } from './utils';
+export { flattenTree } from './utils/flattenTree';
+export { getAllDescendantIds } from './utils/getAllDescendantIds';
+export { getPath } from './utils/getPath';
 export { VirtualizedTree } from './VirtualizedTree';
 export type { VirtualizedTreeProps, VirtualizedTreeRef } from './VirtualizedTree';
