+++ src/HeadlessTree/utils/getPath.ts
@@ -1,4 +1,4 @@
-import type { BasicTreeItem, TreeData, TreeItemId } from './types';
+import type { BasicTreeItem, TreeData, TreeItemId } from '../types';
 
 /**
  * @description Find the path to a specific node in the tree using iterative DFS
@@ -16,7 +16,10 @@ import type { BasicTreeItem, TreeData, TreeItemId } from './types';
  * ```
  * @returns ['1', '2', '3']
  */
-export const getPath = <T extends BasicTreeItem>(tree: TreeData<T>, targetId: TreeItemId): TreeItemId[] => {
+export const getPath = <CustomData extends BasicTreeItem>(
+  tree: TreeData<CustomData>,
+  targetId: TreeItemId
+): TreeItemId[] => {
   // Stack to store [nodeId, pathToNode] pairs
   const stack: Array<[TreeItemId, TreeItemId[]]> = [];
 
@@ -55,37 +58,4 @@ export const getPath = <T extends BasicTreeItem>(tree: TreeData<T>, targetId: Tr
 
   // Node not found
   return [];
-};
-
-/**
- * @description Return all descendant IDs of a specific node
- * @returns The array of IDs of the descendants of the target node
- * @example
- * ```ts
- * const tree = {
- *   rootIds: ['1'],
- *   items: {
- *     '1': { id: '1', children: ['2'] },
- *     '2': { id: '2', children: ['3'] },
- *     '3': { id: '3', children: [] },
- *   },
- * };
- * ```
- * @returns ['2', '3']
- */
-export const getAllDescendantIds = <T extends BasicTreeItem>(tree: TreeData<T>, id: TreeItemId): TreeItemId[] => {
-  const descendants: TreeItemId[] = [];
-  const queue = [...(tree.items[id]?.children || [])];
-
-  while (queue.length > 0) {
-    const currentId = queue.shift()!;
-    descendants.push(currentId);
-
-    const children = tree.items[currentId]?.children;
-    if (children) {
-      queue.push(...children);
-    }
-  }
-
-  return descendants;
 };
